+++ MAX30105.h
@@ -128,12 +128,12 @@ class MAX30105 {
   void setup(uint8_t powerLevel = 0x1F, uint8_t sampleAverage = 4, uint8_t ledMode = 3, int sampleRate = 400, int pulseWidth = 411, int adcRange = 4096);
 
   // Low-level I2C communication
-  uint8_t readRegister8(int address, char reg);
-  void writeRegister8(int address, char reg, char value);
+  uint8_t readRegister8(uint8_t address, char reg);
+  void writeRegister8(uint8_t address, char reg, char value);
 
  private:
   MicroBitI2C *_i2cPort; //The generic connection to user's chosen I2C hardware
-  int _i2caddr = 0x57;
+  uint8_t _i2caddr = 0x57;
 
   //activeLEDs is the number of channels turned on, and can be 1 to 3. 2 is common for Red+IR.
   uint8_t activeLEDs; //Gets set during setup. Allows check() to calculate how many bytes to read from FIFO
